% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{as.array.tensor}
\alias{as.array.tensor}
\title{Converts tensor to array}
\usage{
\method{as.array}{tensor}(x, index_order = NULL, ...)
}
\arguments{
\item{x}{A labeled tensor object, created by \code{\link{\%_\%}} or \code{\link[=tensor]{tensor()}}.}

\item{index_order}{An index specification created with \code{\link[=.]{.()}}.
The specification needs to match all the labels occurring in \code{x}.
The label order determines the dimension
order of the resulting array.}

\item{...}{Not used.}
}
\description{
Converts tensor to array by stripping the index labels.
An index label order
needs to be provided so that the array's \code{\link[=dim]{dim()}} order is well defined.
}
\examples{
array(1:8, dim = c(2, 2, 2)) \%_\% .(i, +i, k) |> .a(k)
}
\concept{tensor}
