% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covd.R
\name{covd}
\alias{covd}
\title{Covariant Derivative}
\usage{
covd(x, new_ind, g = NULL, act_on = all())
}
\description{
Calculates the covariant derivative
\eqn{\nabla_\rho a_{\mu_{1} \mu_{2} ...}^{\nu_{1}\nu_{2}...}} with respect
to the Levi Civita connection of any tensor field.
The covariant derivative is a generalization of the directional derivative
that is independent of the coordinate system in use.
The result is a new tensor of one rank higher than
the original tensor rank.
}
\seealso{
Wikipedia: \href{https://en.wikipedia.org/wiki/Covariant_derivative}{Covariant Derivative}
}
