% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R, R/tensr.R
\name{\%_\%}
\alias{\%_\%}
\alias{tensor}
\title{Create a labeled tensor}
\usage{
a \%_\% i

tensor(a, index_names, index_positions)
}
\arguments{
\item{a}{An array or any object that can be coerced to an array via
\code{\link[=as.array]{as.array()}}.}

\item{i}{An index slot label specification created with \code{\link[=.]{.()}}.}

\item{index_names}{A character vector of index names / labels.}

\item{index_positions}{A character vector of index positions with two allowed
values "+" and "-", for "upper" and "lower" position respectively.
The length of \code{index_positions} needs to agree with the length of
\code{index_names}.}
}
\value{
A labeled tensor object of class \code{"tensor"} which is an \code{array}
whose dimensions have attached labels.
}
\description{
Creates a labeled tensor from an array. \verb{\%_\%} and \code{tensor()} serve
the same purpose, but typically usage of \verb{\%_\%} is preferred due to
brevity.
\code{tensor()} is exported to provide a standard-evaluation interface
as well which might be useful under some circumstances.
}
\examples{
a <- array(1:4, dim = c(2, 2))
a \%_\% .(i, j)
}
\concept{create_tensor}
