% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensr_instances.R
\name{d}
\alias{d}
\title{Kronecker delta}
\usage{
d(n)
}
\arguments{
\item{n}{The dimension.}
}
\value{
A function that expects index labels. See \code{\link[=.]{.()}}.
}
\description{
Provides a labeled generalized Kronecker delta.
}
\examples{
d(3)(i, j)

d(3)(i, j, k, l)
}
\seealso{
\code{\link[calculus:delta]{calculus::delta()}}
}
