% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{.}
\alias{.}
\title{Index slot label specification}
\usage{
.(...)
}
\arguments{
\item{...}{Index labels separated by commas optionally prefixed by "+" and "-"
to indicate the index position (upper and lower respectively).
If no prefix is provided, a lower index ("-") is assumed.
This argument uses non-standard evaluation: any R symbol
that is not a reserved keyword can be used.}
}
\value{
A named list of two character vectors representing the
index label names and index position.
}
\description{
This function creates a index slot label specification. Any R symbol
can serve as a label. \code{.()} is typically used in conjunction with
\link{\%_\%}.
}
\examples{
# three lower index slots
.(i,j,k)

# one lower and upper index
.(i, +j)
}
\concept{tensor}
