% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.R
\name{.t}
\alias{.t}
\title{Index transformation}
\usage{
.t(x, ..., g = getOption("ricci.g"))
}
\arguments{
\item{x}{A labeled tensor object, created by \code{\link{\%_\%}} or \code{\link[=tensor]{tensor()}}.}

\item{...}{Any number of transformation specifications separated by a comma.}

\item{g}{An (unlabeled) covariant metric tensor as array used for raising and
lowering indices.}
}
\value{
A modified tensor object.
}
\description{
\code{.t()} allows to substitute index labels, raising and lowering indices
and perform Kronecker products in any number or order. Multiple transformations
are separated by a comma ",".
}
\details{
\itemize{
\item Substituting index label (without altering position)
\verb{<label1> -> <label2>}
\item Raising index:
\verb{-<label1> -> +<label1>}
\item Lowering index:
\verb{+<label1> -> -<label2>}
\item Kronecker product:
\verb{.(<label1>, <label2>, ..., <labeln-1>) -> <labeln+1>}
}
}
\examples{
a <- array(1:8, dim = c(2, 2, 2))

# raising an index and contracting
a \%_\% .(i, j, k) |>
  .t(-j -> +i)

# taking the Kronecker product
a \%_\% .(i, j, k) |>
  .t(.(j, k) -> l)
}
\concept{tensor}
