% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_instances.R
\name{g_eucl_cart}
\alias{g_eucl_cart}
\alias{g_eucl_sph}
\title{Euclidean metric tensor}
\usage{
g_eucl_cart(n)

g_eucl_sph(n)
}
\arguments{
\item{n}{The dimension of the metric tensor.}
}
\value{
A function that expects index labels (see \code{\link[=.]{.()}}) and returns a
labeled array with dimension \code{c(n, n)}.
}
\description{
Provides a labeled Euclidean metric tensor of \eqn{\mathbb{E}^n}{E^n}.
Note that the numeric result will be different depending on the
chosen index positions.
If no labels are specified, an unlabeled n by n
array of the covariant metric tensor is returned.
}
\details{
\code{g_eucl_cart()} returns a numeric (constant) tensor in Cartesian coordinates,
while \code{g_eucl_sph()} returns a symbolic tensor field in generalized spherical
coordinates \eqn{{r, \phi_1, \phi_2, ..., \phi_{n-1}}}{{r, `ph1`, `ph2`, ..., `ph(n-2)`}}.
As usual, spherical coordinates are degenerate at \eqn{\phi_l = 0}, so be
careful around those points.
}
\examples{
g_eucl_cart(3)(i, j)
g_eucl_cart(3)(+i, +j)
g_eucl_cart(3)()
g_eucl_sph(3)(i, j)
g_eucl_sph(3)(+i, +j)
g_eucl_sph(3)()
}
\seealso{
Wikipedia: \href{https://en.wikipedia.org/wiki/Metric_tensor#Euclidean_metric}{Euclidean metric tensor}

Other metric tensors: 
\code{\link{g_mink}()}
}
\concept{metric tensors}
\concept{tensor_instance}
