% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kron.R
\name{kr}
\alias{kr}
\title{Kronecker product}
\usage{
kr(x, ...)
}
\arguments{
\item{x}{A labeled tensor object.}

\item{...}{Index labels separated by commas optionally prefixed by "+" and "-"
to indicate the index position (upper and lower respectively).
If no prefix is provided, a lower index ("-") is assumed.
This argument uses non-standard evaluation, so any R symbol can
be passed.}
}
\value{
A labeled tensor object of class \code{"tensor"} which is an \code{array}
whose dimensions have attached labels.
}
\description{
Kronecker product
}
