% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_ops.R
\name{l}
\alias{l}
\title{Lower tensor indices}
\usage{
l(x, ..., g = NULL)
}
\arguments{
\item{x}{A labeled tensor object, created by \code{\link{\%_\%}} or \code{\link[=tensor]{tensor()}}.}

\item{...}{Any number of index expressions. The indices need to occur
in \code{x}.}

\item{g}{An (unlabeled) covariant metric tensor as array used for raising and
lowering indices.
If no metric tensor is provided, indices are raised/lowered with
the identity matrix.}
}
\value{
A modified tensor object.
}
\description{
\code{l()} lowers specified tensor indices using a covariant metric tensor
provided in \code{g}.
Note that the order of indices is not preserved due to performance reasons.
An error is thrown if the specified indices do not exist or are not in the
correct position.
}
\seealso{
Other tensor operations: 
\code{\link{Ops.tensor}()},
\code{\link{asym}()},
\code{\link{kron}()},
\code{\link{r}()},
\code{\link{subst}()},
\code{\link{sym}()}
}
\concept{tensor operations}
\concept{tensor_ops}
