% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_ops.R
\name{subst}
\alias{subst}
\title{Substitute tensor labels}
\usage{
subst(x, ...)
}
\arguments{
\item{x}{A labeled tensor object, created by \code{\link{\%_\%}} or \code{\link[=tensor]{tensor()}}.}

\item{...}{Any number of expressions (separated by a comma) of the form
\verb{<label1> -> <label2>}.}
}
\value{
A modified tensor object.
}
\description{
Substitutes tensor labels with other labels. This is simply a renaming
procedure. The operation might trigger implicit diagonal subsetting.
An error is thrown if the specified indices do not exist.
}
\seealso{
Other tensor operations: 
\code{\link{asym}()},
\code{\link{kron}()},
\code{\link{l}()},
\code{\link{r}()},
\code{\link{sym}()}
}
\concept{tensor operations}
\concept{tensor_ops}
