% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_instances.R
\name{d}
\alias{d}
\title{Kronecker delta}
\usage{
d(n)
}
\arguments{
\item{n}{The dimension.}
}
\value{
A function that expects index labels (see \code{\link[=.]{.()}}) and returns a
labeled tensor. The underlying data will differs depending on
the number of labels provided.
}
\description{
Provides a labeled generalized Kronecker delta. In the special
case of two labels this represents simply the identity matrix.
}
\examples{
d(3)(i, j)

d(3)(i, j, k, l)
}
\seealso{
Underlying implementation: \code{\link[calculus:delta]{calculus::delta()}}

Other tensor symbols: 
\code{\link{e}()}
}
\concept{tensor symbols}
\concept{tensor_instance}
