% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_instances.R
\name{e}
\alias{e}
\title{Levi Civita epsilon}
\usage{
e(...)
}
\arguments{
\item{...}{Index labels separated by commas optionally prefixed by "+" and "-"
to indicate the index position (upper and lower respectively).
If no prefix is provided, a lower index ("-") is assumed.
This argument uses non-standard evaluation, so any R symbol can
be passed.}
}
\description{
Provides a labeled Levi Civita epsilon tensor.
}
\examples{
e(i, j)

e(i, j, k)
}
\seealso{
\code{\link[calculus:epsilon]{calculus::epsilon()}}
}
\concept{tensor_instance}
