% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensr_instances.R
\name{g_eucl}
\alias{g_eucl}
\title{Euclidean metric tensor}
\usage{
g_eucl(n)
}
\arguments{
\item{n}{The dimension.}
}
\value{
A function that expects index labels. See \code{\link[=.]{.()}}.
}
\description{
Provides a labeled Euclidean metric tensor in \code{n} dimensions.
}
\examples{
g_eucl(4)(i, j)

g_eucl(4)(+i, +j)
}
\concept{tensor_instance}
