% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_instances.R
\name{g_mink}
\alias{g_mink}
\title{Minkowski metric tensor}
\usage{
g_mink(n)
}
\arguments{
\item{n}{The dimension.}
}
\value{
A function that expects index labels (see \code{\link[=.]{.()}}) and returns a
labeled tensor.
}
\description{
Provides a labeled Minkowski metric tensor in \code{n} dimensions.
Note that the numeric result will be different depending on the
chosen index positions.
}
\examples{
g_mink(4)(i, j)

g_mink(4)(+i, +j)
}
\seealso{
Other metric tensors: 
\code{\link{g_eucl}()}
}
\concept{metric tensors}
\concept{tensor_instance}
