% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensr_instances.R
\name{g_mink}
\alias{g_mink}
\title{Minkowski metric tensor}
\usage{
g_mink(n)
}
\arguments{
\item{n}{The dimension.}
}
\value{
A function that expects index labels. See \code{\link[=.]{.()}}.
}
\description{
Provides a labeled Minkowski metric tensor in \code{n} dimensions.
}
\examples{
g_mink(4)(i, j)

g_mink(4)(+i, +j)
}
\concept{tensor_instance}
