% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_instances.R
\name{g_mink}
\alias{g_mink}
\title{Minkowski metric tensor}
\usage{
g_mink(n)
}
\arguments{
\item{n}{The dimension of the metric tensor.}
}
\value{
A function that expects index labels (see \code{\link[=.]{.()}}) and returns a
labeled tensor with dimension \code{c(n, n)}.
}
\description{
Provides a labeled Minkowski metric tensor in \code{n} dimensions in
Cartesian coordinates with signature \eqn{(-1, 1, 1, ...)}{`c(-1, 1, 1, ...)`}.
Note that the numeric result will be different depending on the
chosen index positions.
If no labels are specified, an unlabeled n by n
array of the covariant metric tensor is returned.
}
\examples{
g_mink(4)(i, j)
g_mink(4)(+i, +j)
g_mink(4)()
}
\seealso{
Wikipedia \href{https://en.wikipedia.org/wiki/Minkowski_space#Minkowski_metric}{Minkowski metric tensor}

Other metric tensors: 
\code{\link{g_eucl_cart}()}
}
\concept{metric tensors}
\concept{tensor_instance}
